fix a few mistakes and update command handlers
The definition and usage of the pfr_disk_read function was incorrect.
The function did not know what size the type name or detail value would
be, because I passed an uninitialized copy of the size field instead of a
pointer to it.

Added temporary command handler for show. This should just use a list
function with a blank filter whenever I add that.